onReady for de1

+++ lib/src/models/device/impl/de1/de1.subscriptions.dart
@@ -74,7 +74,7 @@ extension De1Subscriptions on De1 {
       targetPressure: setGroupPressure,
       targetFlow: setGroupFlow,
       profileFrame: frameNumber,
-      steamTemperature: steamTemp.toDouble(),
+      steamTemperature: steamTemp,
     );
     _snapshotStream.add(_currentSnapshot);
   }
